move withdraw_parts
